Basic game added.
Grid based. Each grid square will map to a certain object, currently only wall, empty or player. Player co-ordinates are stored separately. Game loop continues until a quit is performed.

A clock is also implemented to keep a frames per second.

At this point, the aim of the game is for the player to collect randomly generated targets, the score increasing as more are collected.